harmonize return values of imgc_parse_*

+++ lzo.h
@@ -1,4 +1,4 @@
 #include <stdint.h>
 #include <sys/types.h>
 
-int lzo_decompress(const uint8_t *buf, size_t len, uint8_t *out, size_t outlen);
+size_t lzo_decompress(const uint8_t *buf, size_t len, uint8_t *out, size_t outlen);
